+++ functions/wordpress/auth/registerUser.js
@@ -1,5 +1,5 @@
+import mutationRegisterUser from '@/lib/wordpress/auth/mutationRegisterUser'
 import {initializeWpApollo} from '@/lib/wordpress/connector'
-import mutationRegisterUser from '@/lib/wordpress/users/mutationRegisterUser'
 
 /**
  * Register a user in WP.
