+++ lib/wordpress/auth/mutationLoginUser.js
