+++ lib/wordpress/auth/mutationRegisterUser.js
